requirements.txt
Django==5.2.6
djangorestframework==3.16.1
dotenv==0.9.9
pillow==11.3.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.2.0
django-filter==23.3
setuptools==65.0.0